+++ src/components/SettingsScopePage/index.ts
@@ -1,0 +1,3 @@
+export { SettingsScopePage } from './SettingsScopePage';
+export * from './types';
+export * from './scopes';